Move isShadowIn() in AST
