+++ src/refactorings/inline-variable-or-function/inline-variable.ts
@@ -253,7 +253,7 @@ class InlinableIdentifier implements InlinableCode {
       enter(node, ancestors) {
         if (!ast.isSelectableNode(node)) return;
         if (!ast.areEqual(self.id, node)) return;
-        if (isShadowIn(self.id, ancestors)) return;
+        if (ast.isShadowIn(self.id, ancestors)) return;
 
         const selection = Selection.fromAST(node.loc);
         const isSameIdentifier = selection.isInsideNode(self.id);
@@ -291,40 +291,6 @@ interface IdentifierToReplace {
   loc: ast.SourceLocation;
   isInUnaryExpression: boolean;
   shorthandKey: string | null;
-}
-
-function isShadowIn(
-  id: ast.Identifier,
-  ancestors: ast.TraversalAncestors
-): boolean {
-  // A variable is "shadow" if one of its ancestor redefines the Identifier.
-  return ancestors.some(
-    ({ node }) => isDeclaredInFunction(node) || isDeclaredInScope(node)
-  );
-
-  function isDeclaredInFunction(node: ast.Node): boolean {
-    return (
-      ast.isFunctionDeclaration(node) &&
-      node.params.some(node => ast.areEqual(id, node))
-    );
-  }
-
-  function isDeclaredInScope(node: ast.Node): boolean {
-    return (
-      ast.isBlockStatement(node) &&
-      node.body.some(
-        child =>
-          ast.isVariableDeclaration(child) &&
-          child.declarations.some(
-            declaration =>
-              ast.isVariableDeclarator(declaration) &&
-              ast.areEqual(id, declaration.id) &&
-              // Of course, if it's the inlined variable it's not a shadow!
-              declaration.id !== id
-          )
-      )
-    );
-  }
 }
 
 // 📦 Composites
